Moved domain classes to common project
