+++ build.gradle
@@ -6,6 +6,7 @@ import org.apache.tools.ant.filters.ReplaceTokens
 
 buildscript {
     repositories {
+        mavenLocal()
         mavenCentral()
         jcenter()
     }
@@ -28,6 +29,7 @@ plugins {
 apply plugin: 'io.github.robwin.jgitflow'
 
 repositories {
+    mavenLocal()
     mavenCentral()
     maven { url "https://airdev.jfrog.io/airdev/libs-releases/" }
 }
@@ -50,6 +52,7 @@ dependencies {
     compileOnly 'com.google.code.findbugs:annotations:3.0.1'
     compile 'org.hibernate:hibernate-java8:5.0.12.Final'
     compile 'org.opentestsystem.ap:ap-common:0.3.26'
+    compile 'org.opentestsystem.ap:ap-imrt-common:0.0.1-SNAPSHOT'
     compile 'org.springframework.boot:spring-boot-starter-web'
     compile 'org.springframework.boot:spring-boot-starter-actuator'
     compile 'org.springframework.cloud:spring-cloud-starter-config'
