+++ src/main/java/org/opentestsystem/ap/imrt/iis/ItemIngestServiceApplication.java
@@ -2,6 +2,7 @@ package org.opentestsystem.ap.imrt.iis;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
@@ -9,6 +10,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @SpringBootApplication
 @EnableAsync
 @EnableJpaAuditing
+@EntityScan({"org.opentestsystem.ap.imrt.common"})
 @EnableTransactionManagement
 public class ItemIngestServiceApplication {
