--- src/main/java/org/opentestsystem/ap/imrt/iis/model/ImrtItem.java
@@ -1,136 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import org.hibernate.annotations.Type;
-import org.hibernate.validator.constraints.NotBlank;
-import org.opentestsystem.ap.common.model.Item;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Basic;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
-import java.time.Instant;
-
-@Entity
-@Table(name = "item")
-@EntityListeners(AuditingEntityListener.class)
-public class ImrtItem extends BaseEntity {
-
-    @Id
-    @Column
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer key;
-
-    @NotBlank
-    private String id;
-
-    private String subject="";
-
-    private String grade="";
-
-    @NotBlank
-    private String workflowStatus;
-
-    @NotBlank
-    private String itemType;
-
-    @Column(name = "dok")
-    private String depthOfKnowledge="";
-
-    @NotBlank
-    private String itemCreatedBy;
-
-    @Basic
-    private Instant itemCreatedAt;
-
-    @Type(type = "jsonb")
-    @Column(columnDefinition = "jsonb")
-    @NotNull
-    private Item itemJson;
-
-    public Integer getKey() {
-        return key;
-    }
-
-    public void setKey(Integer key) {
-        this.key = key;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getSubject() {
-        return subject;
-    }
-
-    public void setSubject(String subject) {
-        this.subject = subject;
-    }
-
-    public String getGrade() {
-        return grade;
-    }
-
-    public void setGrade(String grade) {
-        this.grade = grade;
-    }
-
-    public String getWorkflowStatus() {
-        return workflowStatus;
-    }
-
-    public void setWorkflowStatus(String workflowStatus) {
-        this.workflowStatus = workflowStatus;
-    }
-
-    public String getItemType() {
-        return itemType;
-    }
-
-    public void setItemType(String itemType) {
-        this.itemType = itemType;
-    }
-
-    public String getDepthOfKnowledge() {
-        return depthOfKnowledge;
-    }
-
-    public void setDepthOfKnowledge(String depthOfKnowledge) {
-        this.depthOfKnowledge = depthOfKnowledge;
-    }
-
-    public String getItemCreatedBy() {
-        return itemCreatedBy;
-    }
-
-    public void setItemCreatedBy(String itemCreatedBy) {
-        this.itemCreatedBy = itemCreatedBy;
-    }
-
-    public Instant getItemCreatedAt() {
-        return itemCreatedAt;
-    }
-
-    public void setItemCreatedAt(Instant itemCreatedAt) {
-        this.itemCreatedAt = itemCreatedAt;
-    }
-
-    public Item getItemJson() {
-        return itemJson;
-    }
-
-    public void setItemJson(Item itemJson) {
-        this.itemJson = itemJson;
-    }
-
-}
