--- src/main/java/org/opentestsystem/ap/imrt/iis/model/ImrtItemBuilder.java
@@ -1,99 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import org.opentestsystem.ap.common.model.Item;
-
-import java.time.Instant;
-
-/**
- * Builder for ImrtItem objects
- */
-public class ImrtItemBuilder {
-
-    private Integer key;
-    private String id;
-    private String subject="";
-    private String grade="";
-    private String workflowStatus;
-    private String itemType;
-    private String depthOfKnowledge="";
-    private Instant itemCreatedAt;
-    private String itemCreatedBy;
-    private Item itemJson;
-    private String updatedBy;
-
-    public ImrtItemBuilder withKey(int key) {
-        this.key = key;
-        return this;
-    }
-
-    public ImrtItemBuilder withId(String id) {
-        this.id = id;
-        return this;
-    }
-
-    public ImrtItemBuilder withSubject(String subject) {
-        this.subject = subject;
-        return this;
-    }
-
-    public ImrtItemBuilder withGrade(String grade) {
-        this.grade = grade;
-        return this;
-    }
-
-    public ImrtItemBuilder withWorkflowStatus(String workflowStatus) {
-        this.workflowStatus = workflowStatus;
-        return this;
-    }
-
-    public ImrtItemBuilder withItemType(String itemType) {
-        this.itemType = itemType;
-        return this;
-    }
-
-    public ImrtItemBuilder withDepthOfKnowledge(String depthOfKnowledge) {
-        this.depthOfKnowledge = depthOfKnowledge;
-        return this;
-    }
-
-    public ImrtItemBuilder withItemCreatedBy(String itemCreatedBy) {
-        this.itemCreatedBy = itemCreatedBy;
-        return this;
-    }
-
-    public ImrtItemBuilder withItemCreatedAt(Instant itemCreatedAt) {
-        this.itemCreatedAt = itemCreatedAt;
-        return this;
-    }
-
-    public ImrtItemBuilder withItem(Item item) {
-        this.itemJson = item;
-        return this;
-    }
-
-    public ImrtItemBuilder withUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-        return this;
-    }
-
-    /**
-     * Generate a fully populated ImrtItem
-     * @return  A fully populated ImrtItem
-     */
-    public ImrtItem build() {
-        ImrtItem imrtItem = new ImrtItem();
-        imrtItem.setKey(key);
-        imrtItem.setId(id);
-        imrtItem.setSubject(subject);
-        imrtItem.setGrade(grade);
-        imrtItem.setWorkflowStatus(workflowStatus);
-        imrtItem.setItemType(itemType);
-        imrtItem.setDepthOfKnowledge(depthOfKnowledge);
-        imrtItem.setItemCreatedAt(itemCreatedAt);
-        imrtItem.setItemCreatedBy(itemCreatedBy);
-        imrtItem.setItemJson(itemJson);
-        imrtItem.setUpdatedBy(updatedBy);
-        return imrtItem;
-    }
-
-}
