--- src/main/java/org/opentestsystem/ap/imrt/iis/model/ItemGitInformation.java
@@ -1,59 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import org.hibernate.validator.constraints.NotBlank;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "item_git")
-@EntityListeners(AuditingEntityListener.class)
-public class ItemGitInformation extends BaseEntity {
-    @OneToOne
-    private ImrtItem item;
-
-    @Id
-    private Integer projectId;
-
-    @NotBlank
-    private String projectPath;
-
-    @NotBlank
-    private String currentCommitHash;
-
-    public ImrtItem getItem() {
-        return item;
-    }
-
-    public void setItem(ImrtItem item) {
-        this.item = item;
-    }
-
-    public Integer getProjectId() {
-        return projectId;
-    }
-
-    public void setProjectId(Integer projectId) {
-        this.projectId = projectId;
-    }
-
-    public String getProjectPath() {
-        return projectPath;
-    }
-
-    public void setProjectPath(String projectPath) {
-        this.projectPath = projectPath;
-    }
-
-    public String getCurrentCommitHash() {
-        return currentCommitHash;
-    }
-
-    public void setCurrentCommitHash(String currentCommitHash) {
-        this.currentCommitHash = currentCommitHash;
-    }
-}
