--- src/main/java/org/opentestsystem/ap/imrt/iis/model/ItemGitInformationBuilder.java
@@ -1,79 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-/**
- * Builder for ItemGitInformation objects
- */
-public class ItemGitInformationBuilder {
-
-    private Integer projectId;
-    private ImrtItem imrtItem;
-    private String projectPath;
-    private String currentCommitHash;
-    private String updatedBy;
-
-    /**
-     * Sets the projectId
-     * @param projectId  unique projectId
-     * @return  a reference to this object
-     */
-    public ItemGitInformationBuilder withProjectId(Integer projectId) {
-        this.projectId = projectId;
-        return this;
-    }
-
-    /**
-     * Sets the imrtItem
-     * @param imrtItem the imrtItem to set
-     * @return a reference to this project
-     */
-    public ItemGitInformationBuilder withImrtItem(ImrtItem imrtItem) {
-        this.imrtItem = imrtItem;
-        return this;
-    }
-
-    /**
-     * Sets the projectPath
-     * @param projectPath the project path
-     * @return a reference to this object
-     */
-    public ItemGitInformationBuilder withProjectPath(String projectPath) {
-        this.projectPath = projectPath;
-        return this;
-    }
-
-    /**
-     * Sets the current commit hash
-     * @param currentCommitHash  the current commit hash
-     * @return a reference to this object
-     */
-    public ItemGitInformationBuilder withCurrentCommitHash(String currentCommitHash) {
-        this.currentCommitHash = currentCommitHash;
-        return this;
-    }
-
-    /**
-     * Sets the current updatedBy value
-     * @param updatedBy  the updatedBy value
-     * @return a reference to this object
-     */
-    public ItemGitInformationBuilder withUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-        return this;
-    }
-
-    /**
-     * Generate a fully populated ItemGitInformation
-     * @return  A fully populated ItemGitInformation object
-     */
-    public ItemGitInformation build() {
-        ItemGitInformation itemGit = new ItemGitInformation();
-
-        itemGit.setItem(imrtItem);
-        itemGit.setProjectId(projectId);
-        itemGit.setProjectPath(projectPath);
-        itemGit.setCurrentCommitHash(currentCommitHash);
-        itemGit.setUpdatedBy(updatedBy);
-
-        return itemGit;
-    }
-}
