--- src/main/java/org/opentestsystem/ap/imrt/iis/model/ItemLog.java
@@ -1,37 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Basic;
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.Table;
-import java.time.Instant;
-
-@Entity
-@Table(name="item_log")
-@EntityListeners(AuditingEntityListener.class)
-public class ItemLog extends BaseEntity {
-    @EmbeddedId
-    private ItemLogKey itemLogKey;
-
-    @Basic
-    private Instant commitDate;
-
-    public ItemLogKey getItemLogKey() {
-        return itemLogKey;
-    }
-
-    public void setItemLogKey(ItemLogKey itemLogKey) {
-        this.itemLogKey = itemLogKey;
-    }
-
-    public Instant getCommitDate() {
-        return commitDate;
-    }
-
-    public void setCommitDate(Instant commitDate) {
-        this.commitDate = commitDate;
-    }
-}
