--- src/main/java/org/opentestsystem/ap/imrt/iis/model/ItemLogKey.java
@@ -1,38 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
-import javax.persistence.Embeddable;
-import javax.persistence.ManyToOne;
-import javax.validation.constraints.NotNull;
-import java.io.Serializable;
-
-/**
- * Database Key for item_log table
- */
-@SuppressFBWarnings("SE_BAD_FIELD") // Required because Item object, contained inside ImrtItem, does not implement Serializabl
-@Embeddable
-public class ItemLogKey implements Serializable {
-
-    @NotNull
-    @ManyToOne
-    private ImrtItem item;
-
-    @NotNull
-    private String commitHash;
-
-    public ItemLogKey() {}
-
-    public ItemLogKey(ImrtItem item, String commitHash) {
-        this.item = item;
-        this.commitHash = commitHash;
-    }
-
-    public ImrtItem getItem() {
-        return item;
-    }
-
-    public String getCommitHash() {
-        return commitHash;
-    }
-}
