+++ src/main/java/org/opentestsystem/ap/imrt/iis/model/ItemRevision.java
@@ -1,5 +1,9 @@
 package org.opentestsystem.ap.imrt.iis.model;
 
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.StandardId;
+
 import java.util.List;
 import java.util.Optional;
