+++ src/main/java/org/opentestsystem/ap/imrt/iis/model/ItemRevisionImpl.java
@@ -5,6 +5,12 @@ import org.opentestsystem.ap.common.model.AbstractAssessmentItemCore;
 import org.opentestsystem.ap.common.model.Item;
 import org.opentestsystem.ap.common.model.ItemMetadata;
 import org.opentestsystem.ap.common.model.StimItemCore;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ImrtItemBuilder;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformationBuilder;
+import org.opentestsystem.ap.imrt.common.model.StandardId;
+import org.opentestsystem.ap.imrt.common.model.StandardIdBuilder;
 
 import java.time.Instant;
 import java.util.List;
