--- src/main/java/org/opentestsystem/ap/imrt/iis/model/StandardId.java
@@ -1,74 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="standard_id")
-@EntityListeners(AuditingEntityListener.class)
-public class StandardId extends BaseEntity {
-    @EmbeddedId
-    private StandardIdKey standardIdKey;
-
-    private String claim;
-
-    private String contentDomain;
-
-    private String target;
-
-    private String emphasis;
-
-    private String commonCoreStandard;
-
-    public StandardIdKey getStandardIdKey() {
-        return standardIdKey;
-    }
-
-    public void setStandardIdKey(StandardIdKey standardIdKey) {
-        this.standardIdKey = standardIdKey;
-    }
-
-    public String getClaim() {
-        return claim;
-    }
-
-    public void setClaim(String claim) {
-        this.claim = claim;
-    }
-
-    public String getContentDomain() {
-        return contentDomain;
-    }
-
-    public void setContentDomain(String contentDomain) {
-        this.contentDomain = contentDomain;
-    }
-
-    public String getTarget() {
-        return target;
-    }
-
-    public void setTarget(String target) {
-        this.target = target;
-    }
-
-    public String getEmphasis() {
-        return emphasis;
-    }
-
-    public void setEmphasis(String emphasis) {
-        this.emphasis = emphasis;
-    }
-
-    public String getCommonCoreStandard() {
-        return commonCoreStandard;
-    }
-
-    public void setCommonCoreStandard(String commonCoreStandard) {
-        this.commonCoreStandard = commonCoreStandard;
-    }
-}
