--- src/main/java/org/opentestsystem/ap/imrt/iis/model/StandardIdBuilder.java
@@ -1,109 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-public class StandardIdBuilder {
-    private ImrtItem imrtItem;
-    private Integer type;
-    private String claim;
-    private String contentDomain;
-    private String target;
-    private String commonCoreStandard;
-    private String emphasis;
-    private String updatedBy;
-
-    /**
-     * Sets the imrtItem
-     * @param imrtItem  the imrtItem
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withImrtItem(ImrtItem imrtItem) {
-        this.imrtItem = imrtItem;
-        return this;
-    }
-
-    /**
-     * Sets the type
-     * @param type  the type value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withType(int type) {
-        this.type = type;
-        return this;
-    }
-
-    /**
-     * Sets the claim
-     * @param claim  the claim value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withClaim(String claim) {
-        this.claim = claim;
-        return this;
-    }
-
-    /**
-     * Sets the contentDomain
-     * @param contentDomain  the contentDomain value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withContentDomain(String contentDomain) {
-        this.contentDomain = contentDomain;
-        return this;
-    }
-
-    /**
-     * Sets the target
-     * @param target  the target value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withTarget(String target) {
-        this.target = target;
-        return this;
-    }
-
-    /**
-     * Sets the emphasis
-     * @param emphasis  the emphasis value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withEmphasis(String emphasis) {
-        this.emphasis = emphasis;
-        return this;
-    }
-
-    /**
-     * Sets the commonCoreStandard
-     * @param commonCoreStandard  the commonCoreStandard value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withCommonCoreStandard(String commonCoreStandard) {
-        this.commonCoreStandard = commonCoreStandard;
-        return this;
-    }
-
-    /**
-     * Sets the updatedBy
-     * @param updatedBy  the updatedBy value
-     * @return a reference to this object
-     */
-    public StandardIdBuilder withUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-        return this;
-    }
-
-    /**
-     * Generate a StandardId
-     * @return a StandardId
-     */
-    public StandardId build() {
-        StandardIdKey key = new StandardIdKey(imrtItem, type);
-        StandardId id = new StandardId();
-        id.setStandardIdKey(key);
-        id.setClaim(claim);
-        id.setContentDomain(contentDomain);
-        id.setTarget(target);
-        id.setEmphasis(emphasis);
-        id.setCommonCoreStandard(commonCoreStandard);
-        id.setUpdatedBy(updatedBy);
-        return id;
-    }
-}
