--- src/main/java/org/opentestsystem/ap/imrt/iis/model/StandardIdKey.java
@@ -1,38 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
-import javax.persistence.Embeddable;
-import javax.persistence.ManyToOne;
-import javax.validation.constraints.NotNull;
-import java.io.Serializable;
-
-/**
- * Database Key for standard_id table
- */
-@SuppressFBWarnings("SE_BAD_FIELD") // Required because Item object, contained inside ImrtItem, does not implement Serializabl
-@Embeddable
-public class StandardIdKey implements Serializable {
-
-    @NotNull
-    @ManyToOne
-    private ImrtItem item;
-
-    @NotNull
-    private Integer type;
-
-    public StandardIdKey() {}
-
-    public StandardIdKey(ImrtItem item, Integer type) {
-        this.item = item;
-        this.type = type;
-    }
-
-    public ImrtItem getItem() {
-        return item;
-    }
-
-    public Integer getType() {
-        return type;
-    }
-}
