--- src/main/java/org/opentestsystem/ap/imrt/iis/model/StimLink.java
@@ -1,24 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="stim_link")
-@EntityListeners(AuditingEntityListener.class)
-public class StimLink extends BaseEntity {
-    @EmbeddedId
-    private StimLinkKey stimLinkKey;
-
-    public StimLinkKey getStimLinkKey() {
-        return stimLinkKey;
-    }
-
-    public void setStimLinkKey(StimLinkKey stimLinkKey) {
-        this.stimLinkKey = stimLinkKey;
-    }
-}
