--- src/main/java/org/opentestsystem/ap/imrt/iis/model/StimLinkBuilder.java
@@ -1,49 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-public class StimLinkBuilder {
-    private ImrtItem imrtItem;
-    private ImrtItem stimImrtItem;
-    private String updatedBy;
-
-    /**
-     * Sets the imrtItem
-     * @param imrtItem  the imrtItem
-     * @return a reference to this object
-     */
-    public StimLinkBuilder withImrtItem(ImrtItem imrtItem) {
-        this.imrtItem = imrtItem;
-        return this;
-    }
-
-    /**
-     * Sets the stimImrtItem
-     * @param stimImrtItem  the stimImrtItem
-     * @return a reference to this object
-     */
-    public StimLinkBuilder withStimImrtItem(ImrtItem stimImrtItem) {
-        this.stimImrtItem = stimImrtItem;
-        return this;
-    }
-
-    /**
-     * Sets updatedBy
-     * @param updatedBy  the updatedBy value
-     * @return a reference to this object
-     */
-    public StimLinkBuilder withUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-        return this;
-    }
-
-    /**
-     * Generate a StimLink
-     * @return a StimLink
-     */
-    public StimLink build() {
-        StimLinkKey key = new StimLinkKey(imrtItem, stimImrtItem);
-        StimLink stimLink = new StimLink();
-        stimLink.setStimLinkKey(key);
-        stimLink.setUpdatedBy(updatedBy);
-        return stimLink;
-    }
-}
