--- src/main/java/org/opentestsystem/ap/imrt/iis/model/StimLinkKey.java
@@ -1,42 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.model;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
-import javax.persistence.Embeddable;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.validation.constraints.NotNull;
-import java.io.Serializable;
-
-/**
- * Database Key for standard_id table
- */
-@SuppressFBWarnings("SE_BAD_FIELD")  // Required because Item object, contained inside ImrtItem, does not implement Serializable
-@Embeddable
-public class StimLinkKey implements Serializable {
-    @NotNull
-    @ManyToOne
-    private ImrtItem item;
-
-    @NotNull
-    @ManyToOne
-    @JoinColumn(name = "item_key_stim")
-    private ImrtItem itemStim;
-
-    public  StimLinkKey() {
-
-    }
-
-    public StimLinkKey(ImrtItem item, ImrtItem itemStim) {
-        this.item = item;
-        this.itemStim = itemStim;
-    }
-
-    public ImrtItem getItem() {
-        return item;
-    }
-
-    public ImrtItem getItemStim() {
-        return itemStim;
-    }
-}
