+++ src/main/java/org/opentestsystem/ap/imrt/iis/repository/ImrtItemRepository.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
