+++ src/main/java/org/opentestsystem/ap/imrt/iis/repository/ItemLogRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
-import org.opentestsystem.ap.imrt.iis.model.ItemLog;
-import org.opentestsystem.ap.imrt.iis.model.ItemLogKey;
+import org.opentestsystem.ap.imrt.common.model.ItemLog;
+import org.opentestsystem.ap.imrt.common.model.ItemLogKey;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
