+++ src/main/java/org/opentestsystem/ap/imrt/iis/repository/StandardIdRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
-import org.opentestsystem.ap.imrt.iis.model.StandardId;
-import org.opentestsystem.ap.imrt.iis.model.StandardIdKey;
+import org.opentestsystem.ap.imrt.common.model.StandardId;
+import org.opentestsystem.ap.imrt.common.model.StandardIdKey;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
