+++ src/main/java/org/opentestsystem/ap/imrt/iis/repository/StimLinkRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
-import org.opentestsystem.ap.imrt.iis.model.StimLink;
-import org.opentestsystem.ap.imrt.iis.model.StimLinkKey;
+import org.opentestsystem.ap.imrt.common.model.StimLink;
+import org.opentestsystem.ap.imrt.common.model.StimLinkKey;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
