+++ src/main/java/org/opentestsystem/ap/imrt/iis/service/DomainManager.java
@@ -1,10 +1,10 @@
 package org.opentestsystem.ap.imrt.iis.service;
 
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.StimLink;
+import org.opentestsystem.ap.imrt.common.model.StimLinkBuilder;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
-import org.opentestsystem.ap.imrt.iis.model.StimLink;
-import org.opentestsystem.ap.imrt.iis.model.StimLinkBuilder;
 import org.opentestsystem.ap.imrt.iis.repository.ImrtItemRepository;
 import org.opentestsystem.ap.imrt.iis.repository.ItemGitInformationRepository;
 import org.opentestsystem.ap.imrt.iis.repository.StandardIdRepository;
