+++ src/main/java/org/opentestsystem/ap/imrt/iis/service/ItemUpdateService.java
@@ -2,8 +2,8 @@ package org.opentestsystem.ap.imrt.iis.service;
 
 import org.gitlab4j.api.models.RepositoryFile;
 import org.opentestsystem.ap.common.model.Item;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
 import org.opentestsystem.ap.imrt.iis.client.ItemBankClient;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformation;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevisionImpl;
 import org.opentestsystem.ap.imrt.iis.repository.ItemGitInformationRepository;
@@ -36,18 +36,18 @@ public class ItemUpdateService {
 
     private final ProjectLockService projectLockService;
     private final ItemBankClient itemBankClient;
-    private final ItemParser itemUtil;
+    private final ItemParser itemParser;
     private final DomainManager domainManager;
     private final ItemGitInformationRepository itemGitInformationRepository;
     private final OperationalEventService operationalEventService;
 
     public ItemUpdateService(ProjectLockService projectLockService, ItemBankClient itemBankClient,
-                             ItemParser itemUtil, DomainManager domainManager,
+                             ItemParser itemParser, DomainManager domainManager,
                              ItemGitInformationRepository itemGitInformationRepository,
                              OperationalEventService operationalEventService) {
         this.projectLockService = projectLockService;
         this.itemBankClient = itemBankClient;
-        this.itemUtil = itemUtil;
+        this.itemParser = itemParser;
         this.domainManager = domainManager;
         this.itemGitInformationRepository = itemGitInformationRepository;
         this.operationalEventService = operationalEventService;
@@ -76,9 +76,7 @@ public class ItemUpdateService {
             // We got the lock, so process the update
             performItemUpdate(itemId);
         } finally {
-            if (lock.isPresent()) {
-                projectLockService.unlockProject(itemId, lock.get());
-            }
+            lock.ifPresent(aLong -> projectLockService.unlockProject(itemId, aLong));
         }
     }
 
@@ -97,7 +95,7 @@ public class ItemUpdateService {
                 byte json[] = Base64.getDecoder().decode(itemFile.get().getContent());
 
                 // This will throw a runtime exception if there is a problem converting the item
-                Item item = itemUtil.readItem(json);
+                Item item = itemParser.readItem(json);
 
                 // See if we have any existing itemGitInformation for this item in the database
                 ItemGitInformation existingGitInfo = itemGitInformationRepository.findOne(itemId);
