+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/ImrtItemRepositoryTest.java
@@ -2,8 +2,8 @@ package org.opentestsystem.ap.imrt.iis.repository;
 
 import org.junit.Test;
 import org.opentestsystem.ap.common.model.StimItem;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
 import org.opentestsystem.ap.imrt.iis.BaseIntegrationTest;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.time.Instant;
