+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/ItemGitInformationRepositoryTest.java
@@ -2,10 +2,10 @@ package org.opentestsystem.ap.imrt.iis.repository;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformationBuilder;
 import org.opentestsystem.ap.imrt.iis.BaseIntegrationTest;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformation;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformationBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.time.Instant;
