+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/ItemLogBuilder.java
@@ -1,8 +1,8 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemLog;
-import org.opentestsystem.ap.imrt.iis.model.ItemLogKey;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemLog;
+import org.opentestsystem.ap.imrt.common.model.ItemLogKey;
 
 import java.time.Instant;
