+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/ItemLogRepositoryTest.java
@@ -2,9 +2,9 @@ package org.opentestsystem.ap.imrt.iis.repository;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemLog;
 import org.opentestsystem.ap.imrt.iis.BaseIntegrationTest;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemLog;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.time.Instant;
