+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/StimLinkRespositoryTest.java
@@ -2,10 +2,10 @@ package org.opentestsystem.ap.imrt.iis.repository;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.StimLink;
+import org.opentestsystem.ap.imrt.common.model.StimLinkBuilder;
 import org.opentestsystem.ap.imrt.iis.BaseIntegrationTest;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.StimLink;
-import org.opentestsystem.ap.imrt.iis.model.StimLinkBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import static org.assertj.core.api.Assertions.assertThat;
