+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/TestImrtItemBuilder.java
@@ -1,8 +1,8 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
 import org.opentestsystem.ap.common.model.StimItem;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItemBuilder;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ImrtItemBuilder;
 
 import java.time.Instant;
