+++ src/test/java/org/opentestsystem/ap/imrt/iis/repository/TestStandardIdBuilder.java
@@ -1,8 +1,8 @@
 package org.opentestsystem.ap.imrt.iis.repository;
 
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.StandardId;
-import org.opentestsystem.ap.imrt.iis.model.StandardIdKey;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.StandardId;
+import org.opentestsystem.ap.imrt.common.model.StandardIdKey;
 
 public class TestStandardIdBuilder {
     private ImrtItem imrtItem;
