+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/DomainManagerTest.java
@@ -8,10 +8,10 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.StimLink;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
-import org.opentestsystem.ap.imrt.iis.model.StimLink;
 import org.opentestsystem.ap.imrt.iis.repository.ImrtItemRepository;
 import org.opentestsystem.ap.imrt.iis.repository.ItemGitInformationRepository;
 import org.opentestsystem.ap.imrt.iis.repository.StandardIdRepository;
