+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/ItemRevisionTest.java
@@ -9,11 +9,11 @@ import org.opentestsystem.ap.common.model.AssessmentItemCore;
 import org.opentestsystem.ap.common.model.Item;
 import org.opentestsystem.ap.common.model.ItemMetadata;
 import org.opentestsystem.ap.common.model.StimItemCore;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformation;
-import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
-import org.opentestsystem.ap.imrt.iis.model.ItemRevisionImpl;
-import org.opentestsystem.ap.imrt.iis.model.StandardId;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
+import org.opentestsystem.ap.imrt.common.model.ItemRevision;
+import org.opentestsystem.ap.imrt.common.model.ItemRevisionImpl;
+import org.opentestsystem.ap.imrt.common.model.StandardId;
 
 import java.io.IOException;
 import java.net.URL;
