+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/ItemUpdateServiceTest.java
@@ -9,9 +9,9 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.opentestsystem.ap.common.model.Item;
+import org.opentestsystem.ap.imrt.common.model.ImrtItem;
+import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
 import org.opentestsystem.ap.imrt.iis.client.ItemBankClient;
-import org.opentestsystem.ap.imrt.iis.model.ImrtItem;
-import org.opentestsystem.ap.imrt.iis.model.ItemGitInformation;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevisionImpl;
 import org.opentestsystem.ap.imrt.iis.repository.ItemGitInformationRepository;
 import org.opentestsystem.ap.imrt.iis.util.ItemParser;
